64-bit
git-svn-id: http://svn.netlabs.org/repos/kbuild/trunk@793 8c291193-6614-0410-a08d-adc6e07170bb

+++ src/kDepIDB/kDepIDB.c
@@ -42,7 +42,7 @@
 #endif
 #include "kDep.h"
 
-#define OFFSETOF(type, member)  ( (int)(void *)&( ((type *)(void *)0)->member) )
+#define OFFSETOF(type, member)  ( (int)(size_t)(void *)&( ((type *)(void *)0)->member) )
 
 /*#define DEBUG*/
 #ifdef DEBUG
